ncurses: Disable stripping during `make install'
ncurses uses `install -s' to strip binary files during installation,
which uses `strip' in PATH. We can not set PATH to the cross
toolchain at this stage, since we are dealing with both cross
compilation and host compilation at the same time.

Using `install -s', which in turn uses `strip' that can not process
foreign binaries. Disable stripping for now.